+++ packages/@expo/cli/src/start/startAsync.ts
@@ -105,7 +105,7 @@ export async function startAsync(
     await devServerManager.bootstrapTypeScriptAsync();
   }
 
-  if (!settings.webOnly && !options.devClient) {
+  if (!env.EXPO_NO_DEPENDENCY_VALIDATION && !settings.webOnly && !options.devClient) {
     await profile(validateDependenciesVersionsAsync)(projectRoot, exp, pkg);
   }